Refactor top bar styling
