+++ frontend/src/components/HomeTopBar/HomeTopBar.module.css
@@ -1,14 +1,10 @@
-.topBar {
-  width: 100%;
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  padding-top: 0.5%;
-}
-
 .title {
   display: inline-block;
   padding-left: 15px;
   font-size: 25px;
   color: #24305e;
+}
+
+.spacer {
+  padding-right: 20px;
 }
